% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_PaRGHA.R
\name{auto_PaRGHA}
\alias{auto_PaRGHA}
\title{Cria tabelas e salva no banco de dados}
\usage{
auto_PaRGHA(
  db_name,
  host,
  port,
  user,
  password,
  year_start,
  month_start,
  year_end,
  month_end,
  cloud = FALSE,
  save_Rdata_to_path = "./",
  save_path = tempdir()
)
}
\arguments{
\item{db_name}{String. Nome do banco de dados.}

\item{host}{String. Host.}

\item{port}{String. Porta de conexão.}

\item{user}{String. Nome do usuario que tem acesso ao banco}

\item{password}{String. Senha do usuario para acessa o banco}

\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{cloud}{Logical. O padrao é FALSE. Se TRUE, salva as tabelas no DW; caso contrario, cria um RData com todas as tabelas}

\item{save_Rdata_to_path}{String.Caminho para o diretorio onde deve ser salvo o arquivo RData que contém todas as tabelas do BI. O padrao e o diretorio atual "./".}

\item{save_path}{String. Diretorio onde os arquivos DBC serao salvos. O padrao é tempdir().}
}
\value{
Se cloud = FALSE, a função salva um arquivo .RData com o nome tabelas_do_Painel no diretório especificado pelo argumento save_Rdata_to_path.
}
\description{
Cria as tabelas de Leitos, Habilitacoes, Equipamentos e do Numerador/Denominador dos indicadores, e salva cada tabela no banco de dados remoto
}
\examples{
\dontrun{
  auto_PaRGHA(
    db_name = "postgres",
    host = "teste.postgres.database.azure.com",
    port = "1234",
    user = "gustavo",
    password = "-",
    year_start = 2021,
    month_start = 1,
    year_end = 2021,
    month_end = 2,
    cloud = FALSE,
    save_Rdata_to_path = "./"
  )
}

}
