% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure}
\alias{azure}
\title{Obtem ...}
\usage{
azure(data, tabela)
}
\arguments{
\item{data}{DataFrame. Dados que serao inseridos na tabela.}

\item{tabela}{String. Nome da tabela do banco (leitos, habilitacao, equipamentos,numerador_denominador).}
}
\value{
...
}
\description{
...
}
\examples{
\dontrun{
  azure()
}

}
