% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Organizes DBC files to be downloaded into groups}
\usage{
chunk(files_name, data_type = NULL)
}
\arguments{
\item{files_name}{Um vetor contendo os nomes dos arquivos.}

\item{data_type}{String. O padrão é NULL. Valor aceito "PA" ou "SP".}
}
\value{
files_chunks, uma lista de vetores de strings.
}
\description{
Obtém os nomes dos arquivos DBC em `files_name` e os divide em grupos (chunks). O número de grupos e a quantidade de arquivos em cada grupo dependem do número de meses especificado.
}
\examples{
\dontrun{chunk( files_name = files_name)}

}
