% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_fast.R
\name{chunk_fast}
\alias{chunk_fast}
\title{Allocate files into groups with a size limit.}
\usage{
chunk_fast(files_name)
}
\arguments{
\item{files_name}{Um vetor contendo os nomes dos arquivos.}
}
\value{
files_chunks, uma lista de vetores de strings.
}
\description{
Uma heurística de empacotamento para alocar arquivos em grupos com limite de tamanho.
}
\examples{
\dontrun{chunk_fast( files_name = files_name)}

}
