% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_raw.R
\name{create_data_raw}
\alias{create_data_raw}
\title{Cria a base de dados será utilizada pelos indicadores}
\usage{
create_data_raw(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  save_path = tempdir()
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{state_abbr}{string ou vetor de strings. Sigla da Unidade Federativa}

\item{save_path}{String. Diretório onde os arquivos DBC serão salvos. O padrão é tempdir().}
}
\value{
Um DataFrame com os dados de RD, RJ e SP
}
\description{
Realiza o download dos arquivos DBC ('SIH-RD', 'SIH-RJ' e 'SIH-SP'), combina as três bases de dados em uma única estrutura, seleciona apenas as variáveis relevantes, cria novas variáveis e retorna uma base de dados contendo exclusivamente as variáveis que serão utilizadas nos cálculos dos indicadores
}
\examples{
\dontrun{
  create_data_raw(
    year_start = 2024,
    month_start = 1,
    year_end = 2024,
    month_end = 6,
    state_abbr = "CE",
    save_path = ".\\\\"
 )
}

}
