% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indicador.R
\name{create_indicador}
\alias{create_indicador}
\title{Faz o calculo dos indicadores}
\usage{
create_indicador(data)
}
\arguments{
\item{data}{DataFrame obtido pela função x. Deve conter no mínimo as seguintes variáveis: ESPEC, ANO_CMPT, MES_CMPT, CNES, DIAS_PERM}
}
\value{
um
}
\description{
Cria o indicador Giro de Leito da Produção Hospitalar por HUF Apresentadas para cada estabelecimento de saúde (CNES), agrupados por ano e mês.
}
\examples{
\dontrun{create_indicador(data)}

}
