% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_Azure.R
\name{DB_Azure}
\alias{DB_Azure}
\title{Escreve em um Banco de Dados Remoto}
\usage{
DB_Azure(db_name, host, port, user, password, tabela, data, tab_labels = NULL)
}
\arguments{
\item{db_name}{String. Nome do banco de dados.}

\item{host}{String. Host.}

\item{port}{String. Porta de conexão.}

\item{user}{String. Nome do usuario que tem acesso ao banco}

\item{password}{String. Senha do usuario para acessa o banco}

\item{tabela}{String. Nome da tabela do banco (data_aih, leitos, habilitacao, equipamentos,numerador_denominador).}

\item{data}{DataFrame. Dados que serao inseridos na tabela.}

\item{tab_labels}{String. Nome da tabela de labels. O padao e NULL}
}
\description{
Escreve em uma das tabelas a seguir: (leitos, habilitacao, equipamentos, numerador_denominador)
}
\examples{
\dontrun{
  DB_Azure(
    db_name = "postgres",
    host = "teste-tcc.postgres.database.azure.com",
    port = "5432",
    user = "gustavo",
    password = "-",
    tabela = "leitos",
    data = data,
    tab_labels = NULL
  )
}

}
