% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download, Filtragem e Armazenamento de Dados dos Estabelecimentos EBSERH nos Sistemas SIH}
\usage{
download_data(
  year_start,
  month_start,
  year_end,
  month_end,
  save_path = tempdir()
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{save_path}{String. Diretório onde os arquivos DBC serão salvos. O padrão é ".\\".}
}
\description{
Realiza o download dos arquivos DBC dos sistemas SIH-RD, SIH-RJ e SIH-SP para todos os estados que possuem pelo menos um estabelecimento da EBSERH, considerando os meses especificados pelo usuário. Após o download, os arquivos são carregados e filtrados para selecionar apenas os estabelecimentos da EBSERH e as variáveis relevantes. Em seguida, os dados filtrados são salvos em uma base no diretório indicado por save_path.
}
\examples{
\dontrun{
  download_data(
    year_start = 2023,
    month_start = 1,
    year_end = 2023,
    month_end = 3,
    save_path = tempdir()
  )
}

}
