% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dbc.R
\name{download_dbc}
\alias{download_dbc}
\title{Downloads DBC data from the specified healthcare systems: 'SIH-RD', 'SIH-RJ', or 'SIH-SP'}
\usage{
download_dbc(
  information_system,
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  save_path = tempdir()
)
}
\arguments{
\item{information_system}{String. Valor aceito "SIH-RD", "SIH-RJ" ou "SIH-SP".}

\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{state_abbr}{string ou vetor de strings. Sigla da Unidade Federativa}

\item{save_path}{String. Diretório onde os arquivos DBC serão salvos. O padrão é tempdir().}
}
\description{
Realiza o download de arquivos DBC dos sistemas de saúde especificados ('SIH-RD', 'SIH-RJ' ou 'SIH-SP') e os salva em um diretório definido pelo usuário, criando automaticamente uma pasta chamada 'file_DBC'.
}
\examples{
  download_dbc(
    information_system = "SIH-RJ",
    year_start = 2023,
    month_start = 1,
    year_end = 2023,
    month_end = 3,
    state_abbr = "CE",
    save_path = tempdir()
  )

}
