% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_CNES.R
\name{get_data_CNES}
\alias{get_data_CNES}
\title{Download dos dados do CNES que será utilizada pelos indicadores}
\usage{
get_data_CNES(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr = "all",
  type_data = "LT",
  save_path = tempdir()
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{state_abbr}{string ou vetor de strings. Sigla da Unidade Federativa}

\item{type_data}{string. O padrao e "LT". Valores aceitos ("LT", "HB", "EQ"). O tipo da base de dados do CNES que sera baixado.}

\item{save_path}{String. Diretório onde os arquivos DBC serão salvos. O padrão é tempdir().}
}
\value{
Um DataFrame com os dados dos estabelecimentos de saude
}
\description{
Realiza o download dos dados do CNES para um dos sistemas especificados (LT = Leitos, HB = Habilitações, EQ = Equipamentos) e seleciona apenas as colunas necessárias.
}
\examples{
\dontrun{
  dados = get_data_CNES(
    year_start = 2024,
    month_start = 1,
    year_end = 2024,
    month_end = 2,
    state_abbr = "all",
    type_data = "LT",
    save_path = tempdir()
 )
}

}
