% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habilitacoes.R
\name{habilitacoes}
\alias{habilitacoes}
\title{Cria a base de dados das Habilitacoes}
\usage{
habilitacoes(year_start, month_start, year_end, month_end)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}
}
\value{
Um DataFrame com os dados das habilitacoes
}
\description{
Cria a base de dados das Habilitacoes dos estabelecimentos EBSERH
}
\examples{
\dontrun{
  dados = habilitacoes(
    year_start = 2024,
    month_start = 1,
    year_end = 2024,
    month_end = 1
 )
}

}
