% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_N_cirurgias_apresentadas.R
\name{i_N_cirurgias_apresentadas}
\alias{i_N_cirurgias_apresentadas}
\title{Número de Cirurgias Apresentadas por HUF}
\usage{
i_N_cirurgias_apresentadas(data)
}
\arguments{
\item{data}{DataFrame obtido pela função x. Deve conter no mínimo as seguintes variáveis: CO_GRUPO_PROC, ANO_CMPT, MES_CMPT, CNES, TOTAL_CIRURGIAS_APRESENTADAS.}
}
\value{
Um DataFrame com o número de procedimentos apresentados de cada CNES, agrupado por ano e mês.
}
\description{
Cria o indicador Número total de Cirurgias/Produção Hospitalar Apresentadas para cada estabelecimento de saúde (CNES), agrupados por ano e mês.
}
\examples{
\dontrun{i_N_cirurgias_apresentadas(data)}

}
