% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_N_internacoes_hospitalares.R
\name{i_N_internacoes_hospitalares}
\alias{i_N_internacoes_hospitalares}
\title{Número de Internações Hospitalares por HUF}
\usage{
i_N_internacoes_hospitalares(data)
}
\arguments{
\item{data}{DataFrame obtido pela função x. Deve conter no mínimo as seguintes variáveis: ANO_CMPT, MES_CMPT, CNES, TIPO_AIH}
}
\value{
Um DataFrame com o número de Internações Hospitalares de cada CNES, agrupado por ano e mês.
}
\description{
Cria o indicador Número total de Internações Hospitalar Apresentadas para cada estabelecimento de saúde (CNES), agrupados por ano e mês.
}
\examples{
\dontrun{i_N_internacoes_hospitalares(data)}

}
