% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_taxa_de_partos_cesareos.R
\name{i_taxa_de_partos_cesareos}
\alias{i_taxa_de_partos_cesareos}
\title{Taxa de partos Cesareos por HUF}
\usage{
i_taxa_de_partos_cesareos(data)
}
\arguments{
\item{data}{DataFrame obtido pela função x. Deve conter no mínimo as seguintes variáveis: ANO_CMPT, MES_CMPT, CNES, PROC_REA}
}
\value{
Um DataFrame com a Taxa de partos Cesareos de cada CNES, agrupado por ano e mês.
}
\description{
Cria o indicador Taxa de partos Cesareos apresentados na produção hospitalar para cada estabelecimento de saúde (CNES), agrupados por ano e mês.
}
\examples{
\dontrun{i_taxa_de_partos_cesareos(data)}

}
