% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_tempo_medio_permanencia_cirurgica.R
\name{i_tempo_medio_permanencia_cirurgica}
\alias{i_tempo_medio_permanencia_cirurgica}
\title{Tempo Medio de Permanencia Cirurgica por HUF}
\usage{
i_tempo_medio_permanencia_cirurgica(data)
}
\arguments{
\item{data}{DataFrame obtido pela função x. Deve conter no mínimo as seguintes variáveis: ANO_CMPT, MES_CMPT, CNES, DIAS_PERM, ESPEC}
}
\value{
Um DataFrame com o tempo medio de permanencia cirurgica de cada CNES, agrupado por ano e mês.
}
\description{
Cria o indicador Tempo Médio de Permanência Cirurgica para cada estabelecimento de saúde (CNES), agrupados por ano e mês.
}
\examples{
\dontrun{i_tempo_medio_permanencia_cirurgica(data)}

}
