% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_tempo_medio_permanencia_clinica.R
\name{i_tempo_medio_permanencia_clinica}
\alias{i_tempo_medio_permanencia_clinica}
\title{Tempo Medio de Permanencia Clinica por HUF}
\usage{
i_tempo_medio_permanencia_clinica(data)
}
\arguments{
\item{data}{DataFrame obtido pela função x. Deve conter no mínimo as seguintes variáveis: ANO_CMPT, MES_CMPT, CNES, DIAS_PERM, ESPEC}
}
\value{
Um DataFrame com o tempo medio de permanencia clinica de cada CNES, agrupado por ano e mês.
}
\description{
Cria o indicador  Tempo Médio de Permanência clinica para cada estabelecimento de saúde (CNES), agrupados por ano e mês.
}
\examples{
\dontrun{i_tempo_medio_permanencia_clinica(data)}

}
