% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_tempo_medio_permanencia_hospitalar.R
\name{i_tempo_medio_permanencia_hospitalar}
\alias{i_tempo_medio_permanencia_hospitalar}
\title{Tempo Medio de Permanencia Hospitalar por HUF}
\usage{
i_tempo_medio_permanencia_hospitalar(data)
}
\arguments{
\item{data}{DataFrame obtido pela função x. Deve conter no mínimo as seguintes variáveis: ANO_CMPT, MES_CMPT, CNES, DIAS_PERM, ESPEC}
}
\value{
Um DataFrame com o tempo medio de permanencia hospitalar de cada CNES, agrupado por ano e mês. E as colunas que foram  utilizadas no calculo do indicador.
}
\description{
Cria o indicador  Tempo Médio de Permanência Hospitalar para cada estabelecimento de saúde (CNES), agrupados por ano e mês.
}
\examples{
\dontrun{i_tempo_medio_permanencia_hospitalar(data)}

}
