% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{indicadores}
\alias{indicadores}
\title{Cria os indicadores dos estabelecimentos da EBSERH}
\usage{
indicadores(data_SIH, data_CNES)
}
\arguments{
\item{data_SIH}{DataFrame. Objeto retornado pela funcao get_data()}

\item{data_CNES}{DataFrame. Objeto retornado pela funcao create_data_raw_cnes()}
}
\value{
Um DataFrame com os indicadores
}
\description{
Cria os indicadores dos estabelecimentos da EBSERH com base nos dados do SIH e do CNES passados como parametro.
}
\examples{
\dontrun{
  dados = indicadores(
    data_SIH = data_SIH,
    data_cnes = data_CNES
  )
}

}
