% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_empilhados.R
\name{indicadores_empilhados}
\alias{indicadores_empilhados}
\title{Retorna a base de dados com todos os indicadores empilhados}
\usage{
indicadores_empilhados(dados)
}
\arguments{
\item{dados}{a base de dados criada pela função indicadores()}
}
\value{
Um DataFrame com os indicadores empilhados
}
\description{
Reestrutura a base de dados geradada pela função indicadores(), criando uma base de dados onde os indicadores estão disposto todos em uma unica coluna.
}
\examples{
\dontrun{
  dados_empilhados = indicadores_empilhados(dados)
}

}
