% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitos.R
\name{leitos}
\alias{leitos}
\title{Cria a base de dados dos leitos}
\usage{
leitos(year_start, month_start, year_end, month_end, labels = TRUE)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{labels}{Logical. O padrao é TRUE. Cria colunas com os rotulos das variaveis categoricas.}
}
\value{
Um DataFrame com os dados das leitos
}
\description{
Cria a base de dados dos leitos dos estabelecimentos EBSERH
}
\examples{
\dontrun{
  dados = leitos(
    year_start = 2024,
    month_start = 1,
    year_end = 2024,
    month_end = 1,
    labels = TRUE
 )
}

}
