% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Leitura e filtragem dos dados DBC}
\usage{
load_data(information_system, save_path = tempdir())
}
\arguments{
\item{information_system}{string. Sigla do sistema de informacao. Valores permitidos: SIH-RD, SIH-RJ e SIH-SP.}

\item{save_path}{String. Diretorio onde os arquivos DBC serao salvos. O padrao e tempdir().}
}
\value{
Um DataFrame com os dados filtrados
}
\description{
Le arquivos DBC baixados pela função download_data, suportando um unico tipo de sistema por execucao: 'SIH-RD', 'SIH-RJ' ou 'SIH-SP'. Filtra os registros para incluir apenas estabelecimentos da EBSERH e seleciona as colunas relevantes.
}
\examples{
\dontrun{
  load_data(
    information_system = "SIH-RD",
    save_path = tempdir()
 )
}

}
