% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_TCC.R
\name{auto_TCC}
\alias{auto_TCC}
\title{Cria tabelas e salva no banco de dados}
\usage{
auto_TCC(
  db_name,
  host,
  port,
  user,
  password,
  year_start,
  month_start,
  year_end,
  month_end,
  save_path = tempdir()
)
}
\arguments{
\item{db_name}{String. Nome do banco de dados.}

\item{host}{String. Host.}

\item{port}{String. Porta de conexão.}

\item{user}{String. Nome do usuario que tem acesso ao banco}

\item{password}{String. Senha do usuario para acessa o banco}

\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{save_path}{String. Diretório onde os arquivos DBC serão salvos. O padrão é tempdir().}
}
\value{
Um DataFrame com os indicadores
}
\description{
Cria as tabelas de Leitos, Habilitacoes, Equipamentos e do Numerador/Denominador dos indicadores, e salva cada tabela no banco de dados remoto
}
\examples{
\dontrun{
  auto_TCC(
    db_name = "postgres",
    host = "db-tcc.postgres.database.azure.com",
    port = "5432",
    user = "gustavo",
    password = "-",
    year_start = 2021,
    month_start = 1,
    year_end = 2021,
    month_end = 1
  )
}

}
