% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_numerador_denominador.R
\name{create_numerador_denominador}
\alias{create_numerador_denominador}
\title{Cria os Numerador e Denominador dos Indicadores}
\usage{
create_numerador_denominador(data_SIH, data_CNES)
}
\arguments{
\item{data_SIH}{DataFrame. Objeto retornado pela funcao get_data()}

\item{data_CNES}{DataFrame. Objeto retornado pela funcao create_data_raw_cnes()}
}
\value{
Um DataFrame com os numeradores e denominadores dos indicadores
}
\description{
Cria o numerador e o denominador necessarios para o calculo dos indicadores.
}
\examples{
\dontrun{create_numerador_denominador(data_SIH,data_CNES)}

}
