% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Downloads DBC data from the specified healthcare systems: 'SIA-PA', 'SIH-RD', 'SIH-RJ', or 'SIH-SP'}
\usage{
download_data(
  year_start,
  month_start,
  year_end,
  month_end,
  save_path = tempdir()
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{save_path}{String. Diretório onde os arquivos DBC serão salvos. O padrão é ".\\".}
}
\description{
Realiza o download de arquivos DBC dos sistemas de saúde especificados ('SIA-PA', 'SIH-RD', 'SIH-RJ' ou 'SIH-SP') e os salva em um diretório definido pelo usuário, criando automaticamente uma pasta chamada 'file_DBC'.
}
\examples{
\dontrun{
  download_data(
    year_start = 2023,
    month_start = 1,
    year_end = 2023,
    month_end = 3,
    save_path = tempdir()
  )
}

}
