% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download dos dados do DataSUS}
\usage{
download_data(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  save_path = tempdir()
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{state_abbr}{string ou vetor de strings. Sigla da Unidade Federativa.}

\item{save_path}{String. Diretório onde os arquivos DBC serão salvos. O padrão é ".\\".}
}
\description{
Realiza o download automático dos arquivos DBC dos sistemas SIH-RD, SIH-RJ e SIH-SP, de acordo com os estados especificados em state_abbr e os meses definidos pelo usuário. Os arquivos obtidos são salvos no diretório indicado pelo argumento save_path.
}
\examples{
\dontrun{
  download_data(
    year_start = 2023,
    month_start = 1,
    year_end = 2023,
    month_end = 3,
    state_abbr = "CE",
    save_path = tempdir()
  )
}

}
