% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{indicadores}
\alias{indicadores}
\title{Cria os indicadores dos estabelecimentos da EBSERH}
\usage{
indicadores(data_SIH, data_CNES, labels_CNES = TRUE)
}
\arguments{
\item{data_SIH}{DataFrame. Objeto retornado pela funcao get_data()}

\item{data_CNES}{DataFrame. Objeto retornado pela funcao get_data_CNES(type_data = "LT",...)}

\item{labels_CNES}{Logical. O padrao é TRUE. Criar uma coluna com o nome dos estabelecimentos de saude.}
}
\value{
Um DataFrame com os indicadores
}
\description{
Cria os indicadores dos estabelecimentos da EBSERH com base nos dados do SIH e do CNES-LT passados como parametro.
}
\examples{
\dontrun{
  dados = indicadores(
    data_SIH = data_SIH,
    data_CNES = data_CNES,
    labels_CNES = TRUE
  )
}

}
