% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\title{Adiciona rotulos descritivos a partir de colunas de codigos}
\usage{
labels(dados, type_data)
}
\arguments{
\item{dados}{DataFrame.}

\item{type_data}{String. O tipo da base de dados do CNES que sera baixado. Valores aceitos ("cnes", "leitos", "equipamentos", "habilitacao").}
}
\value{
O DataFrame recebido e as novas colunas.
}
\description{
Cria colunas de rotulos associada a colunas de codigos.
}
\examples{
\dontrun{
  dados = labels(dados, "cnes")
}

}
